+++ chess_cogs/music.py
@@ -1,58 +1,12 @@
-import asyncio
+
 import discord
 import yt_dlp as youtube_dl
-
 from discord.ext import commands
 from queue import Queue
 from discord import app_commands
+from dao.youtube import YTDLSource, create_ytdl
 
-# Suppress noise about console usage from errors
-youtube_dl.utils.bug_reports_message = lambda: ''
-
-ytdl_format_options = {
-    'format': 'bestaudio/best',
-    'outtmpl': '%(extractor)s/%(id)s-%(title)s.%(ext)s',
-    'restrictfilenames': True,
-    'noplaylist': True,
-    'nocheckcertificate': True,
-    'ignoreerrors': True,
-    'logtostderr': False,
-    'quiet': False,
-    'progress': True,
-    'no_warnings': True,
-    'extract_flat': True,
-    'default_search': 'auto',
-    'source_address': '0.0.0.0',  # bind to ipv4 since ipv6 addresses cause issues sometimes
-}
-
-ffmpeg_options = {
-    'options': '-vn',
-}
-
-ytdl = youtube_dl.YoutubeDL(ytdl_format_options)
-
-
-class YTDLSource(discord.PCMVolumeTransformer):
-    def __init__(self, source, *, data, volume):
-        super().__init__(source, volume)
-
-        self.data = data
-
-        self.title = data.get('title')
-        self.url = data.get('url')
-
-    @classmethod
-    async def from_url(cls, url, *, loop=None, stream=False, volume):
-        loop = loop or asyncio.get_event_loop()
-        data = await loop.run_in_executor(None, lambda: ytdl.extract_info(url, download=not stream))
-
-        if 'entries' in data:
-            # take first item from a playlist
-            data = data['entries'][0]
-
-        filename = data['url'] if stream else ytdl.prepare_filename(data)
-        return cls(discord.FFmpegPCMAudio(filename, **ffmpeg_options), data=data, volume=volume)
-
+ytdl = create_ytdl()
 
 class Music(commands.Cog):
     def __init__(self, bot):
@@ -74,19 +28,15 @@ class Music(commands.Cog):
                 await channel.send(f'Queue is empty')
                 print(f'Queue is empty')
 
-
     async def ensure_voice(self, interaction: discord.Interaction) -> bool:
         voice_client: discord.VoiceClient = discord.utils.get(self.bot.voice_clients, guild=interaction.guild)
         if voice_client is None:
             if interaction.user.voice:
                 await interaction.user.voice.channel.connect()
                 print(f'Joined {interaction.user.voice.channel.name} channel')
-                # source = discord.PCMVolumeTransformer(discord.FFmpegPCMAudio("X2Download.app - Alright let's do this (Counter Strike Bot Call) - Sound Effect for editing (320 kbps).mp3"))
-                # voice.play(source)
                 return True
             else:
                 await interaction.response.send_message("You are not connected to a voice channel.")
-                # raise commands.CommandError("Author not connected to a voice channel.")
                 return False
         return True
 
@@ -135,12 +85,10 @@ class Music(commands.Cog):
     @app_commands.command(name='skip')
     async def skip(self, interaction: discord.Interaction):
         """Skips current song and plays next in queue."""
-        # await interaction.response.defer()
         vc: discord.VoiceClient = discord.utils.get(self.bot.voice_clients, guild=interaction.guild)
         if vc.is_playing():
             title = vc.source.title
             vc.stop()   
-            # await interaction.followup.send(f'Skipped: **{vc.source.title}**')
             await interaction.response.send_message(f'Skipped: **{title}**')
             print(f'Skipped: {vc.source.title}')
